worker, stateprocessor: fix tx application inconsistency & in doing so improve the efficiency of the block validation process by pulling priority the list of transactors once
