+++ core/state_processor.go
@@ -18,19 +18,15 @@ package core
 
 import (
 	"fmt"
-	"math/big"
-	"strings"
-
-	"github.com/electroneum/electroneum-sc/accounts/abi"
 	"github.com/electroneum/electroneum-sc/common"
 	"github.com/electroneum/electroneum-sc/consensus"
 	"github.com/electroneum/electroneum-sc/consensus/misc"
-	"github.com/electroneum/electroneum-sc/contracts/prioritytransactors"
 	"github.com/electroneum/electroneum-sc/core/state"
 	"github.com/electroneum/electroneum-sc/core/types"
 	"github.com/electroneum/electroneum-sc/core/vm"
 	"github.com/electroneum/electroneum-sc/crypto"
 	"github.com/electroneum/electroneum-sc/params"
+	"math/big"
 )
 
 // StateProcessor is a basic Processor, which takes care of transitioning
@@ -75,7 +71,7 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 	}
 	blockContext := NewEVMBlockContext(header, p.bc, nil)
 	vmenv := vm.NewEVM(blockContext, vm.TxContext{}, statedb, p.config, cfg)
-	transactors := getPriorityTransactors(blockNumber, p.config, vmenv)
+	transactors := GetPriorityTransactors(blockNumber, p.config, vmenv)
 	// Iterate over and process the individual transactions
 	for i, tx := range block.Transactions() {
 		msg, err := tx.AsMessage(types.MakeSigner(p.config, header.Number), header.BaseFee)
@@ -106,69 +102,8 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 	return receipts, allLogs, *usedGas, nil
 }
 
-func getPriorityTransactors(blockNumber *big.Int, config *params.ChainConfig, evm *vm.EVM) common.PriorityTransactorMap {
-	var (
-		address  = config.GetPriorityTransactorsContractAddress(blockNumber)
-		contract = vm.AccountRef(address)
-		method   = "getTransactors"
-		result   = make(common.PriorityTransactorMap)
-	)
-
-	if address != (common.Address{}) {
-		contractABI, _ := abi.JSON(strings.NewReader(prioritytransactors.ETNPriorityTransactorsInterfaceABI))
-		input, _ := contractABI.Pack(method)
-		output, _, err := evm.StaticCall(contract, address, input, params.MaxGasLimit)
-		if err != nil {
-			return result
-		}
-		unpackResult, err := contractABI.Unpack(method, output)
-		if err != nil {
-			return result
-		}
-		transactorsMeta := abi.ConvertType(unpackResult[0], new([]prioritytransactors.ETNPriorityTransactorsInterfaceTransactorMeta)).(*[]prioritytransactors.ETNPriorityTransactorsInterfaceTransactorMeta)
-		for _, t := range *transactorsMeta {
-			// Only add transactors that are within start/end height range
-			if t.StartHeight <= blockNumber.Uint64() && (t.EndHeight >= blockNumber.Uint64() || t.EndHeight == 0) {
-				result[common.HexToPublicKey(t.PublicKey)] = common.PriorityTransactor{
-					IsGasPriceWaiver: t.IsGasPriceWaiver,
-					EntityName:       t.Name,
-				}
-			}
-		}
-	}
-	return result
-}
-
-func getPriorityTransactorByKey(blockNumber *big.Int, publicKey common.PublicKey, config *params.ChainConfig, evm *vm.EVM) (common.PriorityTransactor, bool) {
-	var (
-		address  = config.GetPriorityTransactorsContractAddress(blockNumber)
-		contract = vm.AccountRef(address)
-		method   = "getTransactorByKey"
-	)
-
-	if address != (common.Address{}) {
-		contractABI, _ := abi.JSON(strings.NewReader(prioritytransactors.ETNPriorityTransactorsInterfaceABI))
-		input, _ := contractABI.Pack(method, publicKey.ToUnprefixedHexString())
-		output, _, err := evm.StaticCall(contract, address, input, params.MaxGasLimit)
-		if err != nil {
-			return common.PriorityTransactor{}, false
-		}
-		unpackResult, err := contractABI.Unpack(method, output)
-		if err != nil {
-			return common.PriorityTransactor{}, false
-		}
-		transactorMeta := abi.ConvertType(unpackResult[0], new(prioritytransactors.ETNPriorityTransactorsInterfaceTransactorMeta)).(*prioritytransactors.ETNPriorityTransactorsInterfaceTransactorMeta)
-		// Only add transactors that are within start/end height range
-		if transactorMeta.StartHeight <= blockNumber.Uint64() && (transactorMeta.EndHeight >= blockNumber.Uint64() || transactorMeta.EndHeight == 0) {
-			return common.PriorityTransactor{
-				EntityName:       transactorMeta.Name,
-				IsGasPriceWaiver: transactorMeta.IsGasPriceWaiver,
-			}, true
-		}
-	}
-	return common.PriorityTransactor{}, false
-}
-
+// Keep this function for applying the transaction only, as opposed to verification and pre checks.
+// Priority transactor checks are done within process() for verifying pre-created blocks and the prior wrapper function of applytransaction() for creating blocks
 func applyTransaction(msg types.Message, config *params.ChainConfig, bc ChainContext, author *common.Address, gp *GasPool, statedb *state.StateDB, blockNumber *big.Int, blockHash common.Hash, tx *types.Transaction, usedGas *uint64, evm *vm.EVM) (*types.Receipt, error) {
 	// Create a new context to be used in the EVM environment.
 	txContext := NewEVMTxContext(msg)
@@ -218,7 +153,7 @@ func applyTransaction(msg types.Message, config *params.ChainConfig, bc ChainCon
 // and uses the input parameters for its environment. It returns the receipt
 // for the transaction, gas used and an error if the transaction failed,
 // indicating the block was invalid.
-func ApplyTransaction(config *params.ChainConfig, bc ChainContext, author *common.Address, gp *GasPool, statedb *state.StateDB, header *types.Header, tx *types.Transaction, usedGas *uint64, cfg vm.Config) (*types.Receipt, error) {
+func ApplyTransaction(config *params.ChainConfig, bc ChainContext, author *common.Address, gp *GasPool, statedb *state.StateDB, header *types.Header, tx *types.Transaction, usedGas *uint64, cfg vm.Config, transactors common.PriorityTransactorMap) (*types.Receipt, error) {
 	msg, err := tx.AsMessage(types.MakeSigner(config, header.Number), header.BaseFee)
 	if err != nil {
 		return nil, err
@@ -226,9 +161,10 @@ func ApplyTransaction(config *params.ChainConfig, bc ChainContext, author *commo
 	// Create a new context to be used in the EVM environment
 	blockContext := NewEVMBlockContext(header, bc, author)
 	vmenv := vm.NewEVM(blockContext, vm.TxContext{}, statedb, config, cfg)
+
 	// Validate priority transaction
 	if tx.Type() == types.PriorityTxType {
-		transactor, found := getPriorityTransactorByKey(header.Number, msg.PrioritySender(), config, vmenv)
+		transactor, found := transactors[msg.PrioritySender()]
 		if !found {
 			return nil, fmt.Errorf("could not apply tx [%v]: %w", tx.Hash().Hex(), errBadPriorityKey)
 		}
